rename file containing createUser function
